rename package resource -> entry
