+++ knowde/complex/entry/label.py
@@ -17,7 +17,7 @@ from neomodel import (
     ZeroOrOne,
 )
 
-from knowde.complex.resource.category.folder import Entry, MFolder, MResource
+from knowde.complex.entry.category.folder import Entry, MFolder, MResource
 from knowde.primitive.user.repo import LUser
