+++ knowde/complex/entry/repo/sync.py
@@ -5,14 +5,14 @@ from datetime import datetime
 from typing import TYPE_CHECKING, Iterable
 
 from knowde.complex.__core__.tree2net import parse2net
-from knowde.complex.resource.nx2db.save import resource_meta
-from knowde.complex.resource.router import ResourceMetas
+from knowde.complex.entry.nx2db.save import resource_meta
+from knowde.complex.entry.router import ResourceMetas
 from knowde.primitive.__core__.timeutil import TZ
 
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from knowde.complex.resource import ResourceMeta
+    from knowde.complex.entry import ResourceMeta
 
 
 def can_parse(p: Path, show_error: bool) -> bool:  # noqa: FBT001
